# Optimize tm formatting