Optimize tm formatting for:

- C-locale and alternative representation.
- ``%z`` format (timezone offset from UTC).