@vitaut 2021-11-17
Thanks for the PR. Could you provide any benchmark results showing the effects of this change?