@phprus 2021-11-17
Benchmark: https://github.com/phprus/fmt-bench/tree/optimize-tm-formatting-5

*BSD, Apple, glibc: Acceleration 10-20 times (``%z``).
This is ``std::put_time`` and ostrstream overhead.