@phprus 2021-11-17
Other formatters (``%EY``, ...) on all OS with C-locale: 10-20 times