@alexezeder 2021-11-17
> cc @alexezeder

I'm not 100% sure why I was mentioned here, but here are the results on my machine. 😏

Currently, there is no option to run benchmarks for PRs in [fmt_bnchmrk_gnrtr](https://github.com/alexezeder/fmt_bnchmrk_gnrtr) (I doubt it would be implemented someday). But the only one runner there uses the same version of GCC - 11.

### GCC 11 + libstdc++ (C++20)

| Benchmark             | old time | new time  |
|-----------------------|----------|-----------|
| FMTFormatter_z        | 261 ns   | 26.3 ns   |
| FMTFormatterCompile_z | 246 ns   | 11.6 ns   |
| FMTFormatter_Y        | 263 ns   | 26.1 ns   |
| FMTFormatterCompile_Y | 244 ns   | 10.5 ns   |

<details>
<summary>Raw output</summary>

```
cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_CXX_STANDARD=20 ..
```

```
$ ./fmt_test_old
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
FMTFormatter_z               261 ns          261 ns      2692446
FMTFormatterCompile_z        246 ns          246 ns      2851517
FMTFormatter_Y               263 ns          263 ns      2662867
FMTFormatterCompile_Y        244 ns          244 ns      2867342
```

```
$ ./fmt_test_new
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
FMTFormatter_z              26.3 ns         26.3 ns     26136296
FMTFormatterCompile_z       11.6 ns         11.6 ns     61292230
FMTFormatter_Y              26.1 ns         26.1 ns     26920284
FMTFormatterCompile_Y       10.5 ns         10.5 ns     67130168
```
</details>


### Clang 13 + libc++ (C++20)

| Benchmark             | old time | new time |
|-----------------------|----------|----------|
| FMTFormatter_z        | 110 ns   | 30.6 ns  |
| FMTFormatterCompile_z | 90.0 ns  | 10.4 ns  |
| FMTFormatter_Y        | 110 ns   | 30.5 ns  |
| FMTFormatterCompile_Y | 88.4 ns  | 12.7 ns  |

<details>
<summary>Raw output</summary>

```
CC='clang' CXX='clang++' CXXFLAGS='-stdlib=libc++' cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_CXX_STANDARD=20 ..
```

```
$ ./fmt_test_old
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
FMTFormatter_z               110 ns          110 ns      6403692
FMTFormatterCompile_z       90.0 ns         90.0 ns      7797814
FMTFormatter_Y               110 ns          110 ns      6353394
FMTFormatterCompile_Y       88.4 ns         88.4 ns      7909662
```

```
$ ./fmt_test_new
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
FMTFormatter_z              30.6 ns         30.6 ns     23115238
FMTFormatterCompile_z       10.4 ns         10.4 ns     67346905
FMTFormatter_Y              30.5 ns         30.4 ns     22898049
FMTFormatterCompile_Y       12.7 ns         12.7 ns     55214092
```
</details>