@phprus 2021-11-17
@alexezeder, I mentioned you since you wrote a comment about performance regression (https://github.com/fmtlib/fmt/commit/1031eedf27fe55fb7e85d7ce47dcbd3ecebb290e#commitcomment-60128058)

This PR improves performance in those cases (C-locale and timezone offset).