@vitaut 2021-11-21
@phprus, thanks for the benchmark. Could you add `strftime` there for comparison?