@phprus 2021-11-21
@vitaut, I will do a performance test (on Windows and with raw ``strftime``, ``std::time_put``) in the coming days.