@phprus 2021-11-21
@vitaut, benchmark updated: https://github.com/phprus/fmt-bench/tree/optimize-tm-formatting-5