@phprus 2021-11-25
Sorry for delay.

UCRT is are official name for Microsoft C Runtime start from Visual Studio 2015 (https://docs.microsoft.com/en-us/cpp/porting/upgrade-your-code-to-the-universal-crt?view=msvc-170)

I replaced the abbreviation UCRT to another version of the name: Universal CRT.